+++ conreality_scripts/lib/src/script.dart
@@ -28,6 +28,21 @@ class Script {
         : "scripts/$path.lua";
   }
 
+  /// The script name or label.
+  ///
+  /// Note: parsing this metadata requires loading the source file.
+  Future<String> get label => _parseMetadata("label");
+
+  /// A short summary of what the script does.
+  ///
+  /// Note: parsing this metadata requires loading the source file.
+  Future<String> get summary => _parseMetadata("summary");
+
+  /// The author of the script.
+  ///
+  /// Note: parsing this metadata requires loading the source file.
+  Future<String> get author => _parseMetadata("author");
+
   /// Loads the binary data from the script's `.luac` bytecode file.
   Future<ByteData> loadBinary() {
     return rootBundle.load(binaryKey);
@@ -36,5 +51,17 @@ class Script {
   /// Loads the source code from the script's `.lua` source file.
   Future<String> loadSource() {
     return rootBundle.loadString(sourceKey);
+  }
+
+  Future<String> _parseMetadata(final String property) async {
+    final pattern = RegExp(r'^\s*([^=\s]+)\s*=\s*"([^"]+)"');
+    final source = await loadSource();
+    for (final line in source.split("\n")) {
+      final match = pattern.firstMatch(line);
+      if (match != null && match.group(1) == property) {
+        return match.group(2);
+      }
+    }
+    return null;
   }
 }