+++ conreality_scripts/CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.1+20190704.1 - 2019-07-04
+### Added
+- `Script#label` getter
+- `Script#summary` getter
+- `Script#author` getter
+
+### Added
 ## 1.0.0+20190704.1 - 2019-07-04
 ### Added
 - `ConrealityScripts.version` getter